need to sequrety

+++ services/storeservice/internal/service/delete.go
@@ -2,16 +2,35 @@ package storeservice
 
 import (
 	"context"
+	"myapp/pkg/errmsg"
 	"myapp/pkg/richerror"
 	"storeapp/internal/param"
 )
 
-func (s *Service) DeleteStore(ctx context.Context, req param.DeleteStoreRequest) error {
+func (s *Service) DeleteStore(ctx context.Context, req param.DeleteStoreRequest) (param.DeleteStoreResponse, error) {
 	const op = "StoreService.DeleteStore"
-	_, err := s.repo.GetStoreByID(ctx, req.StoreID)
+	store, err := s.repo.GetStoreByID(ctx, req.StoreID)
 	if err != nil {
-		return richerror.New(op).WithErr(err)
+		return param.DeleteStoreResponse{
+			Success: false,
+		}, richerror.New(op).WithErr(err)
 	}
+
+	if store.UserID != req.UserID {
+		return param.DeleteStoreResponse{
+				Success: false,
+			}, richerror.New(op).WithKind(richerror.KindForbidden).
+				WithMessage(errmsg.ErrorMsgUserNotAllowed)
+	}
+
 	err = s.repo.DeleteStore(ctx, req.StoreID)
-	return richerror.New(op).WithErr(err)
+	if err != nil {
+		return param.DeleteStoreResponse{
+			Success: false,
+		}, richerror.New(op).WithErr(err)
+	}
+
+	return param.DeleteStoreResponse{
+		Success: true,
+	}, nil
 }
